# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Certificate Automation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SheetFile {
  id: string;
  name: string;
  modifiedTime?: string;
}

export interface SheetListResponse {
  sheets: SheetFile[];
}

export type SheetDataResponseRowsItem = { [key: string]: string };

export interface SheetDataResponse {
  headers: string[];
  rows: SheetDataResponseRowsItem[];
  totalRows: number;
}

export interface SlideTemplate {
  id: string;
  name: string;
  modifiedTime?: string;
  thumbnailUrl?: string;
}

export interface SlideTemplateListResponse {
  templates: SlideTemplate[];
}

export interface PlaceholderResponse {
  placeholders: string[];
}

export interface CreatedFileResponse {
  id: string;
  name: string;
  url: string;
}

export interface CreateSlideTemplateRequest {
  name: string;
}

export interface CreateSheetRequest {
  name: string;
  headers: string[];
}

export type BatchColumnMap = { [key: string]: string };

export type BatchStatus = (typeof BatchStatus)[keyof typeof BatchStatus];

export const BatchStatus = {
  draft: "draft",
  generating: "generating",
  generated: "generated",
  sending: "sending",
  sent: "sent",
  partial: "partial",
} as const;

export interface Batch {
  id: string;
  name: string;
  sheetId: string;
  sheetName: string;
  templateId: string;
  templateName: string;
  columnMap: BatchColumnMap;
  emailColumn: string;
  nameColumn: string;
  emailSubject?: string;
  emailBody?: string;
  status: BatchStatus;
  driveFolderId?: string | null;
  pdfFolderId?: string | null;
  totalCount: number;
  generatedCount: number;
  sentCount: number;
  createdAt: string;
}

export type CertificateStatus =
  (typeof CertificateStatus)[keyof typeof CertificateStatus];

export const CertificateStatus = {
  pending: "pending",
  generated: "generated",
  sent: "sent",
  failed: "failed",
} as const;

export type CertificateRowData = { [key: string]: string };

export interface Certificate {
  id: string;
  batchId: string;
  recipientName: string;
  recipientEmail: string;
  status: CertificateStatus;
  slideFileId?: string;
  slideUrl?: string;
  pdfFileId?: string;
  pdfUrl?: string;
  sentAt?: string;
  errorMessage?: string;
  rowData?: CertificateRowData;
  createdAt: string;
}

export type BatchDetail = Batch & {
  certificates: Certificate[];
};

export interface BatchListResponse {
  batches: Batch[];
}

export type CreateBatchRequestColumnMap = { [key: string]: string };

export interface CreateBatchRequest {
  name: string;
  sheetId: string;
  sheetName: string;
  tabName?: string;
  templateId: string;
  templateName: string;
  columnMap: CreateBatchRequestColumnMap;
  emailColumn: string;
  nameColumn: string;
  emailSubject?: string;
  emailBody?: string;
}

export interface SendBatchRequest {
  emailSubject: string;
  emailBody: string;
}

export interface BatchOperationResponse {
  success: boolean;
  message: string;
  processed?: number;
  failed?: number;
}

export interface CertificateListResponse {
  certificates: Certificate[];
  total: number;
}

export type GetSheetDataParams = {
  tabName?: string;
};

export type ListCertificatesParams = {
  batchId?: string;
  status?: string;
};