# packages/api-zod/src/generated/types/batch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Certificate Automation API
 * OpenAPI spec version: 0.1.0
 */
import type { BatchColumnMap } from "./batchColumnMap";
import type { BatchStatus } from "./batchStatus";

export interface Batch {
  id: number;
  name: string;
  sheetId: string;
  sheetName: string;
  templateId: string;
  templateName: string;
  columnMap: BatchColumnMap;
  emailColumn: string;
  nameColumn: string;
  emailSubject?: string;
  emailBody?: string;
  status: BatchStatus;
  totalCount: number;
  generatedCount: number;
  sentCount: number;
  createdAt: string;
}