# packages/api-zod/src/generated/types/certificateListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Certificate Automation API
 * OpenAPI spec version: 0.1.0
 */
import type { Certificate } from "./certificate";

export interface CertificateListResponse {
  certificates: Certificate[];
  total: number;
}