# packages/api-zod/src/generated/types/placeholderResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Certificate Automation API
 * OpenAPI spec version: 0.1.0
 */

export interface PlaceholderResponse {
  placeholders: string[];
}